refactor(desktop): share the composer's mention plumbing
Extracts the pure halves of the full composer's mention machinery out of
Composer.tsx so a second surface can call them instead of copying them:

- `composer-mention-tokens.ts` — token minting for every chip kind, the
  index adjustment that keeps zero-width tokens aligned as the plain draft
  changes around them, and the markdown serializer. A verbatim move.
- `hash-references.ts` gains `buildHashReferenceOptions`, the local/peer
  merge and dedupe that was inline in Composer's `#` memo, plus
  `describeHashReferenceThread`.
- `HighlightedAutocompleteLabel.tsx` — its own module now that more than
  one surface renders autocomplete rows.

No behaviour change: Composer.tsx imports what it used to define.
